pyproject.toml
[tool.poetry]
name = "shootbox"
version = "0.2"
description = "2D Top-down Sandbox game, written in Python"
authors = ["FireFall <firekonstantin@mail.ru>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pygame-ce = "^2.4.0"
pypresence = "^4.3.0"
orjson = "^3.9.9"
numpy = "^1.26.1"
numba = "^0.58.1"
toml = "^0.10.2"
tomlkit = "^0.12.3"
rtoml = "^0.10.0"
babel = "^2.14.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
nuitka = "^2.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
